pyproject.toml
[project]
name = "ps-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.26.0",
    "photoshop-python-api>=0.24.1",
    "pywin32>=311",
]
